Add files via upload
In this project we are hiding text data in an image using lsb technology. here we encrypting the text data first before hiding it in image at lsb. so we provided two algorithms for encrypting data which can be selected by filling key section. any integer for first algorithm and 0 for second algorithm. Then hidden image will be created. and data can be extracted from it by uploading it in image section and clicking on extract data.